Reorganised the physics engine code
