+++ indigo/indigo/src/main/scala/indigo/physics/simulation/Solved.scala
@@ -1,0 +1,6 @@
+package indigo.physics.simulation
+
+import indigo.shared.datatypes.Vector2
+import indigo.shared.geometry.Vertex
+
+final case class Solved(nextPosition: Vertex, nextVelocity: Vector2)
